Verify Transaction.nonce is no bigger than 64 bits
This is important to do, because nonces larger than 64 bits are invalid.
In order to do this, I had to add native support for converting
BigIntegers to and from ULong. In order to do that, I had to add native
support for converting ByteArray to and from ULong.

+++ Shared/src/commonTest/kotlin/com/conradkramer/wallet/browser/message/EventMessageTests.kt
@@ -1,5 +1,6 @@
 package com.conradkramer.wallet.browser.message
 
+import com.conradkramer.wallet.bigint.BigInteger
 import com.conradkramer.wallet.browser.event.ConnectEvent
 import com.conradkramer.wallet.ethereum.Quantity
 import kotlin.test.Test
@@ -11,7 +12,7 @@ internal class EventMessageTests() : MessageTests() {
         val message = EventMessage(
             20,
             session,
-            ConnectEvent(Quantity(1))
+            ConnectEvent(Quantity(BigInteger.valueOf(1)))
         )
         val jsonMessage = decode(
             """{
